Expose precListLookahead, document pswap's behaviour better
